refactor repost events
