+++ ndk/utils.tsx
@@ -14,6 +14,7 @@ import {
   finishEvent,
   generatePrivateKey,
   type EventTemplate,
+  Kind,
 } from "nostr-tools";
 import axios from "axios";
 import { bech32 } from "@scure/base";
@@ -498,4 +499,23 @@ export const createAppDataEventTemplate = ({
   });
 
   return eventTemplate;
+};
+
+export const createRepostEvent = (ndk: NDK, event: NDKEvent): NDKEvent => {
+  let created_at = Math.floor(Date.now() / 1000);
+  let tags: NDKTag[] = [
+    ["e", event.id, process.env.NEXT_PUBLIC_STEMSTR_RELAY as string],
+    ["p", event.pubkey],
+  ];
+  let kind = 6;
+  if (event.kind !== Kind.Text) {
+    kind = 16;
+    tags.push(["k", `${event.kind}`]);
+  }
+  let repostEvent = new NDKEvent(ndk);
+  repostEvent.kind = kind;
+  repostEvent.created_at = created_at;
+  repostEvent.tags = tags;
+  repostEvent.content = JSON.stringify(event.rawEvent());
+  return repostEvent;
 };
